+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.1.2]
+* Update scroll_to_index package to latest version  to support Flutter 3.0
+* Update examples to Flutter 3.0
+
 ## [0.1.1]
 * Add collapseOthers and collapseOthersAndSnapToTop expansion behaviors